refactor(netxlite): use *Netx for creating UDP sockets
This diff is similar to https://github.com/ooni/probe-cli/commit/07a048ce8e944ba68b3fd55faf6c1de9a9620485
but here we use *Netx to create UDP sockets.

While there, recognize that UDP code needs its own files
and should not live inside quic{,_test}.go.

Part of https://github.com/ooni/probe/issues/2531
