+++ internal/netxlite/udp_test.go
@@ -1,0 +1,9 @@
+package netxlite
+
+import "testing"
+
+func TestNewUDPListener(t *testing.T) {
+	ql := NewUDPListener()
+	qew := ql.(*udpListenerErrWrapper)
+	_ = qew.UDPListener.(*udpListenerStdlib)
+}
